requirements.txt
click==7.1.1
Flask==1.1.2
Flask-WTF==0.14.3
json5==0.9.4
pandas==1.0.3
python-dotenv==0.15.0
numpy==1.18.1